add gems

+++ db/migrate/20230607071334_change_total_to_be_integer_in_orders.rb
@@ -1,6 +1,6 @@
 class ChangeTotalToBeIntegerInOrders < ActiveRecord::Migration[7.0]
   def up
-    change_column :orders, :total, :decimal
+    change_column :orders, :total, :integer
   end
 
   def down
